Publishing Amazon SNS Messages Privately
In this project, you will be working on a hospital project to send reports online and develop a platform so the patients can access the reports via mobile and push notifications. You will
publish the report to an Amazon SNS keeping it secure and private. Your message will be hosted on an EC2 instance within yourAmazon VPC. By publishing the messages privately, you can improve
the message delivery and receipt through Amazon SNS.